pyproject.toml
[tool.poetry]
name = "sudanese-antiquities-hackathon"
version = "0.1.0"
description = ""
authors = ["jdelorimier <james.delorimier@c4ads.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
clip = {git = "https://github.com/openai/CLIP.git"}
torch = "^2.7.0"
pillow = "^11.2.1"
numpy = "^2.2.6"
scikit-learn = "^1.6.1"
packaging = "^25.0"
torchvision = "^0.22.0"
tqdm = "^4.67.1"
ftfy = "^6.3.1"
regex = "^2024.11.6"
faiss-cpu = "^1.11.0"
timm = "^1.0.15"
matplotlib = "^3.10.3"
openai = "^1.82.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
